Revert "style: apply responsive styles to all components"
