+++ index.html
@@ -1,13 +1,13 @@
 <!doctype html>
-<html lang="en" class="min-h-full">
+<html lang="en">
   <head>
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Zaplytic</title>
   </head>
-  <body class="relative bg-gray-900">
-    <div id="root" class="h-full pt-15 text-gray-100 fs--base"></div>
+  <body>
+    <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
