+++ src/components/About.tsx
@@ -1,0 +1,23 @@
+export default function About() {
+  return (
+    <>
+      <h1>About Zaplytic</h1>
+      <div>
+        Zaplytic, a team of few passionate developers building open-source tools
+        and SaaS products to empower developers, businesses, and everyday users.
+        Our mission is to create high-quality, easy-to-use software that solves
+        real-world problems. We believe in the power of open-source and
+        community collaboration.
+        <h3>We Build</h3>
+        <li>
+          Open-Source Tools: open-source tools to help developers be more
+          productive and efficient.
+        </li>
+        <li>
+          SaaS Products: innovative SaaS products for businesses and
+          individuals.
+        </li>
+      </div>
+    </>
+  );
+}
