+++ src/vite-env.d.ts
@@ -1,8 +1,2 @@
 /// <reference types="vite/client" />
 declare module "*.jpg";
-
-interface project {
-    name: string
-    description: string
-    technology: string
-}
